msm: camera: req_mgr: Rearrange state change sequence
Component bind is responsible for platform bus. QUP i2c bus slaves
can come at probe anytime. This operation is corrupting the slave
list for CRM. This change makes sure that CRM is available after
subdev list is initiate first and then only allows the subdev to be
add in the list.

CRs-Fixed: 3034933
Change-Id: Ia08702a8b38f2f2f590021b997c729131e919d2d
Signed-off-by: Jigarkumar Zala <jzala@codeaurora.org>