qcacld-3.0: Kbuild: Enable nested composites on supported QKI kernels
KERNEL_SUPPORTS_NESTED_COMPOSITES := y is used to enable nested
composite support. The nested composites support is available in
all GKI & QKI kernels beginning with
5.10.20, but unfortunately is not available in any upstream kernel.

When a QKI kernel is used the flag is not set explicitly on the
kernel sources. Current logic sets the flag for GKI kernels only.
So extend the logic to set the flag for QKI kernels also.

Change-Id: I96aa36628aaf66f194c14f8883674da205220117
CRs-Fixed: 3041134
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>