qcacld-3.0: Validate 3-port combo p2p+p2p+sta based on fw cap
Check the existing two connection are in below comibation or not
GO + GO
GO + GC
GC + GC
GO + STA
GC + STA
If the existing two connection are in above combination and third
connection comes as GO or GC or STA then check whether firmware
supports p2p-p2p concurrency or not.
if firmware supports p2p-p2p concurrency and third connection is
GO or GC or STA then only allow the third connection as GO or GC
or STA otherwise do not allow third connection as GO or GC or STA
with above combination.

Change-Id: I0c2408232e84438dc5dd8d104b75fdee126aa31f
CRs-Fixed: 3043845
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>