qcacld-3.0: Send RSO stop in STA + STA roaming
In STA+STA roaming case, when primary interface is changed
and roaming is enabled on other interface rso stop will be sent.
But the roam scan mode is not set to 0.
So add check for primary interface change to reset roam scan mode
before sending RSO stop to FW.

Change-Id: Ia880b58a57b3d9b74f3c4e100d4724aac46b67cd
CRs-Fixed: 3042416
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>