msm: camera: isp: Don't clear lists when disabling all events
On overflow, cam_irq_controller_disable_all function is called from top
half handler, which itself is in a loop of all entries of top half list.
Manipulating the list to remove all elements can leave it in an unstable
state and cause the loop to never terminate. Since events are removed
from all lists when unsubscribing from the IRQs, do not clear the top
half list when disabling all events.

CRs-Fixed: 3043489
Change-Id: Idee0a017cdea1dac73e9a7a00f07463afe324983
Signed-off-by: Anand Ravi <ananravi@codeaurora.org>