qcacld-3.0: Add support for STA+GO and CLI+GO for liberal mode
Currently in liberal mode STA+GO and CLI+GO scenario is not
handled, so when sta is there and Go tries to come up then
GO should follow STA or CLI's channel in case of force scc
but as handling is not there so Go doesn't consider scc
channel.

Fix is to add handling of STA+GO and CLI+GO for liberal mode.

Change-Id: I227fb0661c5c3a41b296fd0c954ac87675f28ed6
CRs-Fixed: 3048501
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>