msm: camera: isp: Add support for IFE scratch buffer
Add support to configure IFE scratch buffer for sHDR/FS use-cases.
BW limiter is enabled for an IFE output port to achieve
the desired spread in frame timing. On occasions where there is
no packet in KMD to apply or for a request that has no buffer
for this IFE output port configure scratch buffer.
This will ensure BW limiter comes into affect, and the frame
is spread to acheive the desired FPS.

CRs-Fixed: 3045706
Change-Id: Id723e7c3d53e65cd7824827b12e86502e8bd91f0
Signed-off-by: Karthik Anantha Ram <kartanan@codeaurora.org>