qcacld-3.0: Change the max vdev number to 5
When 32 STAs connected to APUT, crash will happen because of lack
of resource. The max vdev number was decreased by 1 from the value
of 4 because of nan is not supported. Then FW will received the
value of 3 and finally the tid number is not enough. Change the
max value of vdev number to 5 to fix this issue.

Change-Id: Ia212a0b05922530526439206fd70a44f3ce03838
CRs-Fixed: 3064133
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>